Refactor: 유저프로필 수정 리팩토링 (순수한 view를 분리)
