+++ webapp/src/pages/EditUserProfile/EditUserProfileDetail.jsx
@@ -1,12 +1,6 @@
-import React, { useRef, useState } from 'react';
-import PropTypes from 'prop-types';
-import { userEditParser } from 'service/user.parser';
-import TechStackSelectInput from 'components/TechStackSelectInput';
-import TextInput from 'components/Common/TextInput';
-import Button from 'components/Common/Button';
-import SelectInput from 'components/Common/SelectInput';
+import React, { useState } from 'react';
+import { userEditParser, userPostEditParser } from 'service/user.parser';
 import { skillStackParser } from 'service/skillStack.parser';
-import { belongTeamOptions, hopeSessionOption, jobOptions, skillStack } from 'constant';
 import { editUserValidation } from 'service/user.validation';
 import { useToastNotificationAction } from 'contexts/ToastNotification';
 import { notifyNewMessage } from 'contexts/ToastNotification/action';
@@ -14,22 +8,16 @@ import { TOAST_TYPE } from 'contexts/ToastNotification/type';
 import authApi from 'api/auth.api';
 import useForm from 'hooks/useForm';
 import { userEditType } from 'types/user.type';
-import MarkdownEditor from 'components/MarkdownEditor';
-import * as S from './EditUserProfile.style';
+import useAxios from 'hooks/useAxios';
+import userApi from 'api/user.api';
+import useFileUploader from 'hooks/useFileUploader';
+import EditUserProfileView from './EditUserProfile.view';
 
 EditUserProfileDetail.propTypes = {
   targetUser: userEditType,
-  submitCallback: PropTypes.func.isRequired,
-  onChangeFile: PropTypes.func.isRequired,
-  imageFile: PropTypes.string,
 };
 
-export default function EditUserProfileDetail({
-  targetUser,
-  submitCallback,
-  onChangeFile,
-  imageFile,
-}) {
+export default function EditUserProfileDetail({ targetUser }) {
   const notifyDispatch = useToastNotificationAction();
 
   const parsedTargerUserInfo = userEditParser(targetUser);
@@ -46,11 +34,39 @@ export default function EditUserProfileDetail({
     portfolio,
   } = parsedTargerUserInfo;
 
-  const parsedSkillStackOptions = skillStackParser(skillStack);
   const parsedSkillStack = skillStackParser(techSkills);
 
-  const inputFileRef = useRef();
   const [isNicknameDuplicate, setIsNicknameDuplicate] = useState(false);
+
+  // 수정 요청 api hooks
+  const [state, execution, foreceRefetch] = useAxios({
+    axiosInstance: userApi.EDIT_USER_PROFILE,
+    immediate: false,
+  });
+
+  // s3 이미지 업로더 api hooks
+  const { uploadFileOnS3, imageFile, onChangeFile } = useFileUploader({
+    notifyNewMessage,
+    notifyDispatch,
+  });
+
+  const uploadImageFileBeforeSubmit = async (submitData) => {
+    const response = await uploadFileOnS3();
+    if (response) {
+      const { id, path } = response;
+      return { ...submitData, profileImage: path };
+    }
+    return submitData;
+  };
+
+  // 수정 요청
+  const submitCallback = async (submitData) => {
+    const changedProfileImageSubmitData = await uploadImageFileBeforeSubmit(submitData);
+    const parsedSubmitData = userPostEditParser(changedProfileImageSubmitData);
+    await execution({ data: parsedSubmitData });
+    // TODO: 성공시 이동할 페이지 정해서 이동시키기
+    notifyNewMessage(notifyDispatch, '수정 완료!', TOAST_TYPE.Success);
+  };
 
   const {
     inputValues,
@@ -101,142 +117,22 @@ export default function EditUserProfileDetail({
     }
   };
 
-  const onUploadButtonClick = () => {
-    inputFileRef.current.click();
-  };
-
-  const isSkillsValidateError = isTargetSatisfyValidate('techSkills');
-  const isNicknameValidateError = isTargetSatisfyValidate('nickname');
-  const isSloganValidateError = isTargetSatisfyValidate('slogan');
-
-  const canActiveSubmitButton = !satisfyAllValidates && isNicknameDuplicate;
+  const isNickNameSameWithOrigin = nickname === inputValues.nickname;
 
   return (
-    <S.PostContainer>
-      <S.Form onSubmit={submitHandler} id="editUserProfileForm">
-        <S.ProfileImageContainer>
-          {imageFile ? (
-            <S.InputTypeImageHandler htmlFor="profileImage">
-              <S.ImageThunbnail alt="upload" src={URL.createObjectURL(imageFile)} />
-              <Button
-                theme="secondary"
-                customStyle={S.ImageEditButton}
-                onClick={onUploadButtonClick}
-              >
-                수정
-              </Button>
-            </S.InputTypeImageHandler>
-          ) : (
-            <S.InputTypeImageHandler htmlFor="profileImage">
-              <div>
-                <S.PlusSolid />
-              </div>
-            </S.InputTypeImageHandler>
-          )}
-          <S.HiddenInputHandler
-            ref={inputFileRef}
-            id="profileImage"
-            name="profileImage"
-            type="file"
-            accept="image/*"
-            onChange={onChangeFile}
-          />
-        </S.ProfileImageContainer>
-        <S.InfoContainer>
-          <S.DuplicateCheckInput>
-            <TextInput
-              id="checkDuplicateNickname"
-              name="nickname"
-              label="닉네임"
-              placeholder="닉네임"
-              value={inputValues.nickname}
-              onChange={onChangeHandler}
-              isError={isNicknameValidateError}
-              helperText={validateError.nickname}
-            />
-            <Button
-              type="button"
-              htmlFor="checkDuplicateNickname"
-              theme="secondary"
-              customStyle={S.DuplicateCheckButton}
-              disabled={isNicknameValidateError}
-              onClick={onClickCheckDuplicateNickname}
-            >
-              중복확인
-            </Button>
-          </S.DuplicateCheckInput>
-          <TechStackSelectInput
-            name="techSkills"
-            placeholder="기술"
-            label="기술"
-            selectedTechSkills={inputValues.techSkills}
-            techSkillOptions={parsedSkillStackOptions}
-            onChange={onChangeHandlerWithSelect}
-            isError={isSkillsValidateError}
-            helperText={validateError.techSkills}
-          />
-          <TextInput
-            name="slogan"
-            label="슬로건"
-            placeholder="슬로건"
-            value={inputValues.slogan}
-            onChange={onChangeHandler}
-            isError={isSloganValidateError}
-            helperText={validateError.slogan}
-          />
-          <SelectInput
-            name="hopeSession"
-            label="회망 기간"
-            placeHolder="회망 기간"
-            defaultOption={hopeSessionOption[0]}
-            options={hopeSessionOption}
-            value={inputValues.hopeSession}
-            onChange={onChangeHandlerWithSelect}
-          />
-          <SelectInput
-            name="job"
-            label="직업"
-            placeHolder="직업"
-            defaultOption={jobOptions[0]}
-            options={jobOptions}
-            value={inputValues.job}
-            onChange={onChangeHandlerWithSelect}
-          />
-          <SelectInput
-            name="belongTeam"
-            label="팀 소속 여부"
-            placeHolder="팀 소속 여부"
-            defaultOption={belongTeamOptions[0]}
-            options={belongTeamOptions}
-            value={inputValues.belongTeam.label}
-            onChange={onChangeHandlerWithSelect}
-          />
-          <MarkdownEditor
-            onlyViewer={false}
-            label="자기 소개"
-            placeholder="자기 소개를 입력해주세요."
-            content={introduction}
-          />
-          <TextInput
-            name="portfolio"
-            label="포트폴리오(url)"
-            placeholder="포트폴리오(url)"
-            value={inputValues.portfolio}
-            onChange={onChangeHandler}
-          />
-        </S.InfoContainer>
-      </S.Form>
-      <S.ButtonContainer>
-        <Button
-          theme="primary"
-          type="submit"
-          form="editUserProfileForm"
-          disabled={canActiveSubmitButton}
-          customStyle={S.SubmitButton}
-        >
-          저장
-        </Button>
-      </S.ButtonContainer>
-    </S.PostContainer>
+    <EditUserProfileView
+      inputValues={inputValues}
+      onChangeHandler={onChangeHandler}
+      onChangeHandlerWithSelect={onChangeHandlerWithSelect}
+      submitHandler={submitHandler}
+      validateError={validateError}
+      satisfyAllValidates={satisfyAllValidates}
+      isTargetSatisfyValidate={isTargetSatisfyValidate}
+      isNicknameDuplicate={isNicknameDuplicate}
+      isNickNameSameWithOrigin={isNickNameSameWithOrigin}
+      onClickCheckDuplicateNickname={onClickCheckDuplicateNickname}
+      imageFile={imageFile}
+      onChangeFile={onChangeFile}
+    />
   );
 }
